% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_c}
\alias{bootstrap_c}
\title{bootstrap_c}
\usage{
bootstrap_c(x, B, W, V)
}
\arguments{
\item{x}{Covariate vector}

\item{B}{number of permutations}

\item{W}{sum of the main variable and each target}

\item{V}{difference of the main variable and each target}
}
\description{
This function performs permutation test for given covariate x B times
Also takes as input W instead of Y matrix
}
