% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_degree}
\alias{get_degree}
\title{Computes the score statistic}
\usage{
get_degree(x, y, Y)
}
\arguments{
\item{x}{covariate matrix to test}

\item{y}{vector of main variable}

\item{Y}{matrix of the rest of the variables. The row number of Y should match the length of y.}
}
\value{
double: degree statistic
}
\description{
Computes the score statistic
}
