% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_p_from_degree}
\alias{get_p_from_degree}
\title{Computes the small sample correction}
\usage{
get_p_from_degree(y, Y, d, numsim = 5000)
}
\arguments{
\item{y}{main vector to test against}

\item{Y}{matrix of target vectors}

\item{d}{degree statistic}

\item{numsim}{number of simulations for null distribution}
}
\value{
p-value
}
\description{
Computes the small sample correction
}
