% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_score}
\alias{get_score}
\title{get_score
The order of y1 and y2 does not matter.}
\usage{
get_score(x, y1, y2)
}
\arguments{
\item{x}{covariate vector}

\item{y1}{observation of the first variable}

\item{y2}{observation of the second variable}
}
\description{
get_score
The order of y1 and y2 does not matter.
}
