% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_score_wv_c}
\alias{get_score_wv_c}
\title{get_score_wv_c}
\usage{
get_score_wv_c(x, w, v)
}
\arguments{
\item{x}{Covariate vector}

\item{w}{sum of variable1 and variable2}

\item{v}{difference of variable1 and variable2}
}
\description{
This function takes the sum of two variables, w, the difference of them, v, and one dimensional covariate x
and returns the score test statistic
}
