% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_degree}
\alias{get_degree}
\title{get_degree_c}
\usage{
get_degree(x, y, Y)
}
\arguments{
\item{x}{Covariate vector}

\item{y}{Variable of interest}

\item{Y}{All other variables to test with y}
}
\description{
This function returns a sum statistic d for vector y tested with all other variables in matrix Y against covariate x
}
