% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_p_from_degree}
\alias{get_p_from_degree}
\title{Estimate H}
\usage{
get_p_from_degree(y, Y, d, numsim = 5000)
}
\arguments{
\item{y}{vector: main variable of interest}

\item{Y}{Matrix: other variables of interest to measure the correlation with y}

\item{d}{double: computed degree statistic}

\item{numsim}{integer: number of simulations to draw the null distribution}
}
\value{
p value of the degree statistic
}
\description{
Estimate H
}
